Move version checking code to its own class.
